Switch from `if DEBUG` to `$logger.debug`
I made the logger global because `logger` wasn’t found from within the class when I made it local to that file. There should only be one logger in the program, and I don’t expect to ever have to test the output of the logger, so making it global is fine.

I also simplified some logging code by changing an `each` to a `map`.